requirements.txt
# BloomDirect Claims Recovery System
# Python 3.11+

# Core
streamlit==1.43.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
httpx==0.27.2

# Scheduler
apscheduler==3.10.4

# LLM
anthropic==0.49.0

# LangGraph / orchestration
langgraph==0.3.5
langchain-core==0.3.45

# MCP
mcp==1.5.0

# Data
pandas==2.2.3

# Google / Gmail MCP
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.165.0

# Guardrails
pydantic==2.10.6

# Utilities
python-dateutil==2.9.0
pytz==2024.2
plotly
